handle home and profile exception

+++ app/src/main/java/com/example/shopify/authentication/model/pojo/Address.kt
@@ -6,11 +6,11 @@ import com.google.gson.annotations.SerializedName
 class Address {
     @SerializedName("id")
     @Expose
-    var id: Int? = null
+    var id: Long? = null
 
     @SerializedName("customer_id")
     @Expose
-    var customerId: Int? = null
+    var customerId: Long? = null
 
     @SerializedName("first_name")
     @Expose
